Update, 23 03 2021 21 52

+++ Form1.cs
@@ -11,13 +11,19 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using XRayModsManager.res;
 
 namespace XRayModsManager
 {
     public partial class Form1 : Form
     {
+        private FilesController ctrl;
+        private Thread thread;
+
         public Form1()
         {
+            ctrl = new FilesController();
+
             InitializeComponent();
 
             Load += OnLoad;
@@ -33,26 +39,35 @@ namespace XRayModsManager
             }
         }
 
+        // при нажатии на ссылку
         private void linkLabel1_LinkClicked(object sender, LinkLabelLinkClickedEventArgs e)
         {
             Process.Start("https://github.com/kotleni/xray-modsmanager");
         }
 
+        // при нажатии на кнопку "Запустить игру"
         private void start_Click(object sender, EventArgs e)
         {
             LOG("Запуск игры.");
 
             string[] paths = new string[] {
-                System.Environment.CurrentDirectory + "\\bin\\XR_3DA.exe"
+                System.Environment.CurrentDirectory + "\\bin\\XR_3DA.exe",
+                System.Environment.CurrentDirectory + "\\bin\\xrEngine.exe"
             };
 
             for(var i = 0; i < paths.Length; i++)
             {
                 if (File.Exists(paths[i]))
+                {
                     Process.Start(paths[i]);
+                    return;
+                }
             }
+
+            MessageBox.Show("Не удалось найти файл для запуска игры!", "Ошибка");
         }
 
+        // Удалить папку, со всем содержимым
         public static void DeleteDirectory(string target_dir)
         {
             string[] files = Directory.GetFiles(target_dir);
@@ -72,15 +87,17 @@ namespace XRayModsManager
             Directory.Delete(target_dir, false);
         }
 
+        // при нажатии на кнопку "Обновить моды"
         private void update_Click(object sender, EventArgs e)
         {
             DeleteDirectory("gamedata/");
             Directory.CreateDirectory("gamedata/");
 
-            Thread thread = new Thread(updateMods);
+            thread = new Thread(updateMods);
             thread.Start();
         }
 
+        // функция логирования
         private void LOG(string msg)
         {
             File.AppendAllText("xraymm.log", msg + "\n");
@@ -88,7 +105,8 @@ namespace XRayModsManager
             Console.WriteLine(msg);
         }
        
-        bool first = true;
+        bool first = true; // если первая итерация
+        string currentMod = "";
         private void iterDir(string path)
         {
             if (!first) {
@@ -111,11 +129,28 @@ namespace XRayModsManager
             {
                 var clearFilePath = file.Split(new string[] { "\\" }, 4, StringSplitOptions.None)[3];
 
-                LOG(" Копирование файла " + clearFilePath);
-                File.Copy(file, "gamedata\\" + clearFilePath);
+                if (!File.Exists("gamedata\\" + clearFilePath))
+                {
+                    LOG(" Копирование файла " + clearFilePath);
+                    File.Copy(file, "gamedata\\" + clearFilePath);
+
+                    ctrl.addFile(currentMod, clearFilePath);
+                } else
+                { // если файл уже есть
+                    LOG("Ошибка, файл уже есть! " + clearFilePath);
+
+                    var modfile = ctrl.findFile(clearFilePath);
+
+                    this.BeginInvoke((MethodInvoker) delegate {
+                        MessageBox.Show("Конфликт модов, " + modfile.plugin + " и " + currentMod + ".", "Ошибка");
+                    });
+
+                    thread.Abort();
+                }
             }
         }
 
+        // загрузить мод
         private void loadMod(string name)
         {
             first = true;
@@ -123,16 +158,28 @@ namespace XRayModsManager
             iterDir("mods\\" +  name + "\\gamedata\\");
         }
 
+        // обновить моды
         private void updateMods()
         {
             DateTime dt = DateTime.Now;
             LOG("[" + String.Format("{0:s}", dt) + "]");
 
+            var i = 0;
             foreach (string mod in modsList.CheckedItems)
             {
                 LOG("Установка мода " + mod);
+                currentMod = mod;
+
                 loadMod(mod);
+                
+                this.BeginInvoke((MethodInvoker)delegate {
+                    modsList.SetItemChecked(i, false);
+                });
+
+                i++;
             }
+
+        
         }
     }
 }
